requirements.txt
fastapi
openai
pydantic
requests
python-dotenv
uvicorn[standard]

